+++ Soggins/I18n.cs
@@ -102,9 +102,9 @@ internal static class I18n
 
 	// EXE card
 	public static string ExeCardName => "Soggins.EXE";
-	public static string ExeCardText0 => $"Get <c=status>SMUG</c>. Add 1 of 2 <c=cardtrait>discnt, temp</c> <c={SogginsColor}>{SogginsName}</c> cards to your hand.";
-	public static string ExeCardTextA => $"Get 1 <c=status>SMUG</c>. Add 1 of 2 <c=cardtrait>discnt, temp</c> <c={SogginsColor}>{SogginsName}</c> cards to your hand.";
-	public static string ExeCardTextB => $"<c=status>SMUG</c> = 0. Add 1 of 3 <c=cardtrait>discnt, temp</c> <c={SogginsColor}>{SogginsName}</c> cards to your hand.";
+	public static string ExeCardText0 => $"Get <c=status>SMUG</c>. Add 1 of 3 <c=cardtrait>discnt, temp</c> <c={SogginsColor}>{SogginsName}</c> cards to your hand.";
+	public static string ExeCardTextA => $"Get 1 <c=status>SMUG</c>. Add 1 of 3 <c=cardtrait>discnt, temp</c> <c={SogginsColor}>{SogginsName}</c> cards to your hand.";
+	public static string ExeCardTextB => $"<c=status>SMUG</c> = 0. Add 1 of 5 <c=cardtrait>discnt, temp</c> <c={SogginsColor}>{SogginsName}</c> cards to your hand.";
 
 	public static string PaletteTooltip => "<c=B79CE5>SOGGINS PIGMENT</c>\nYour next colorless card also adds a random <c=card>Halfhearted Apology</c> to your hand.";